+++ areas/acr/src/AzureMcp.Acr/Commands/AcrJsonContext.cs
@@ -7,6 +7,7 @@ using AzureMcp.Acr.Commands.Registry;
 namespace AzureMcp.Acr.Commands;
 
 [JsonSerializable(typeof(RegistryListCommand.RegistryListCommandResult))]
+[JsonSerializable(typeof(RegistryRepositoryListCommand.RegistryRepositoryListCommandResult))]
 [JsonSerializable(typeof(Models.AcrRegistryInfo))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
 internal sealed partial class AcrJsonContext : JsonSerializerContext